Add noscrape, secret, divergent.
noscrape - config option to disable /scrape and /stats globally.

secret - if an allowed torrent file is named *.secret.torrent it will be
excluded from general /scrape and /stats reporting (partly implemented).

divergent - if a peer claims an IP other than the one it's reporting,
this will be noted in /stats.